test: extract StubExportReceiptStore.swift from IssueExportTestSupport.swift
Splits the in-memory ExportReceiptStoring test double (~78 lines) out
of IssueExportTestSupport.swift into its own file so that support file
holds only the primary Mock* actors (extraction + export services).

Byte-identical move. Both files retain their test imports.

IssueExportTestSupport.swift: 362 → 284 lines. Tests: 667.
